Move CLI code into index of cli directory
